reorganized package structures for database implementation
