+++ app/src/main/java/com/vordead/snoozeloo/MainActivity.kt
@@ -5,7 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.vordead.snoozeloo.ui.SnoozelooApp
+import com.vordead.snoozeloo.core.presentation.SnoozelooApp
 import com.vordead.snoozeloo.ui.theme.SnoozelooTheme
 import dagger.hilt.android.AndroidEntryPoint
