+++ app/src/main/java/com/vordead/snoozeloo/alarm/data/local/AlarmEntity.kt
@@ -1,0 +1,2 @@
+package com.vordead.snoozeloo.alarm.data.local
+
