+++ app/src/main/java/com/vordead/snoozeloo/alarm/domain/Alarm.kt
@@ -1,0 +1,2 @@
+package com.vordead.snoozeloo.alarm.domain
+
